xen-boot-cmd: Add xen variable as addendum
For xen boot the u-boot script environment variables needs to
be complied as part of u-boot-xlnx-scr recipe. In order to duplicate
the variables definition for xen and non-xen boot make xen variables
as addendum to u-boot-xlnx-scr recipe.

Upstream-Status: Submitted [https://lists.yoctoproject.org/g/meta-virtualization/message/8233]

Signed-off-by: Sandeep Gundlupet Raju <sandeep.gundlupet-raju@amd.com>
Signed-off-by: Mark Hatle <mark.hatle@amd.com>

Transfer from 2024.2 fork of meta-virtualization, as this won't be in meta-virtualization
scarthgap.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>