meta-xilinx-mali400: linux-xlnx-udev-rules: Always apply the rule
If this layer is enabled, always add the mali400 udev rule.  If run on a
system without a mali400 it will have no effect other then taking a very
small amount of disk space.  This avoids the need to have a MACHINE_ARCH
package when mali400 is enabled in MACHINE_FEATURES.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>