u-boot-xlnx: Add preboot command
Using the preboot command for image recovery as we are using bif with
ddr load addresses for kernel and rootfs.
boot.scr need to be copied to ospi at 0x7f80000 for versal and it will
increase the .bin file size to 132MB, so removing boot.scr and using
preboot config to run the boot command.

Signed-off-by: Raju Kumar Pothuraju <rajukumar.pothuraju@amd.com>
Signed-off-by: Mark Hatle <mark.hatle@amd.com>