meta-xilinx-imgrcvry: Add base multiconfig
Add a reference multiconfig that should enable the the special image-recovery
to be buildable from a multiconfig.  This currently assumes a 'versal'
target due to the hard coded paths.

Note, the paths are hardcoded because normally we append to 'TMPDIR', however
the TMPDIR has already been modified because we may be running from within
a multiconfig.

In order to use this, you would also need a corresponding setting in a
global configuration file, such as:

 BBMULTICONFIG += "xilinx-image-recovery"

 IMGRCRY_DEPENDS = ""
 IMGRCRY_MCDEPENDS = "mc::xilinx-image-recovery:image-recovery-tiny-initramfs:do_image_complete"
 IMGRCRY_DEPLOY_DIR = "${TMPDIR}-${MACHINE}-xilinx-image-recovery/deploy/images/${MACHINE}"
 IMGRCRY_IMAGE_NAME = "image-recovery-tiny-initramfs-${MACHINE}.imgrcry"

Signed-off-by: Mark Hatle <mark.hatle@amd.com>