meta-xilinx-core: embeddedsw: Add workaround for mcdepends
Multiconfig dependencies are configured to be specific to a single from
and to multiconfig, for example:

  mc:default:xilinx-image-recovery:core-image-tiny-initramfs:do_image_complete

The above says, when in the 'default' multiconfig, we depend on the
'xilinx-image-recovery' multiconfig, and need to run the do_image_complete
task within the core-image-tiny-initramfs recipe.

Two workarouns are present, the first adjusts:

  mc::to_mc:recipe:task

to be

  mc:${BB_CURRENT_MC}:to_mc:recipe_task

This will allow the mcdependencies produced by gen-machine-conf to wor in
any multiconfig the user may have.  In the future we will adjust
gen-machine-conf to use ${BB_CURRET_MC} instead.

Second there is a bug/oversight in bitbake where the BB_CURRENT_MC reports
'default', but the system is expecting '' (blank) when in the default Linux
multiconfig.  We will do this conversion to avoid any issues.  A patch for
bitbake is in progress.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>