meta-xilinx-core: qemuboot-xilinx.bbclass: Cleanup variables
QB_MACHINE is now set in each base machine.  This ensures that the default
configuration is correct, and can still be overriden if necessary.
This solves a problem where bitbake -e could fail to expand properly
in some cases.

QB_DEFAULT_KERNEL was removed, and restored back to the default settings.
The defaults now work properly with all tested configurations.
Note: simpleImage.mb is obsolete and no longer used in this context.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>