meta-xilinx-core: qemu-multiarch-helper: Support -dtb as -hw-dtb analog for APU
To ensure it is easier to override the APU device-tree, support -dtb as an
analog for -hw-dtb to the APU.  Additionally, runqemu needs to define a kernel
to pass -dtb, so ignore the -kernel parameter as it doesn't do anything in
the system emulated APU.

Also check and error if an APU DTB is not passed in, either via -dtb or -hw-dtb.
This can cause qemu to lock up, try to prevent this error.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>