iperf3: backport upstream patch
The team working on the XXV + EOE implementation (25G PL Ethernet using SFP+)
observed a bug in iperf3 which causes a roughly 50% drop in the bitrate (from
9.57Gbps down to 5.35Gbps). On scarthgap iperf3 is at version 3.16; this bug
is fixed in version 3.18. Instead of bumping the entire code base to 3.18 it
would be safer to simply apply the fix for this bug onto 3.16 instead.

Link: https://github.com/esnet/iperf/pull/1708/commits/ac6b9f7fd335ddebc5212eed40083ef4cd3cb86d
Signed-off-by: Trevor Woerner <trevor.woerner@amd.com>
Signed-off-by: Mark Hatle <mark.hatle@amd.com>