meta-xilinx-core: qemu-system-aarch64-multiarch: Refactor to support multiple archs
Add signal handling for microblze emulation processes.  Needed if the
APU shuts down unexpectedly.  Note, when APU shuts down correctly, the
microblaze will return a non-zero code.  Ignore this as it leads to a
potentially confusing error.

Add -bootbin arch:file support, as we need a different 'arch' configuration
for future processors.

Much of the above was implemented by:
    Sandeep Gundlupet Raju <sandeep.gundlupet-raju@amd.com>

Signed-off-by: Mark Hatle <mark.hatle@amd.com>