qemu-xilinx-8.2: Add .git suffix to download URLs
Add the .git suffix to the download URLs of
+ berkeley-softfloat-3
+ berkeley-testfloat-3
+ libslirp

to match the way it is specified in Qemu itself.

This also helps work around an issue with populating the local
downloads cache from a shared mirror where for libslirp, only the
files ending in .git/.git.done are copied instead of the ones
without the suffix as expected by this recipe prior to this
change.

Signed-off-by: Christian Kohn <chris.kohn@amd.com>
Signed-off-by: Mark Hatle <mark.hatle@amd.com>