meta-xilinx-standalone(-sdt): Move recipes to correct dir
All of the embeddedsw recipes should be in a directory at the correct
depth but are currently all in recipes-drivers and recipes-libraries.

This change moves all of the recipes into an "embeddedsw" directory at
the correct depth and makes the reauired changes to the two layer
configs and a number of indludes.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>