wic: split out partition label from gpt name
A partition's label is stored as part of a partition's meta-data. Whether or
not labeling is supported depends on the filesystem type and whether or not
that type supports labels. GPT partition names, on the other hand, are stored
in the GPT table's metadata and only depends on whether or not GPT
partitioning is used; support for GPT partition names is not required
per-filesystem.

In a wks configuration file, the user specifies they want a filesystem label
applied to a filesystem by using the '--label <name>' directive. A user
specifies they want a GPT partition name applied to a partition using the
'--part-name <name>' directive.

The wic code conflates these two and sets a partition's GPT name in either
case when a user specifies '--part-name <name>' (correct) and '--label <name>'
(not correct).

Signed-off-by: Trevor Woerner <trevor.woerner@amd.com>
Signed-off-by: Mark Hatle <mark.hatle@amd.com>