util-linux: carry the version from upsream master
wic uses the sfdisk(-native) utility from the util-linux project. Adding
support for 4k sectors requires a version of sfdisk that can handle
variable-sized sectors specified on the command-line. The version of sfdisk in
scarthgap (2.39.3) does not have this feature, but the version in master
(2.40.4) does.

Signed-off-by: Trevor Woerner <trevor.woerner@amd.com>

The microblaze util-linux patch needs to be updated for 2.40 as well.

Move existing integration to specifically target the Poky version,
while a new 2.40 version is created.  This should allow either version
if the user decided to use PREFERRED_VERSION.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>