wic: add support for sector size to sfdisk
Recent versions of the 'sfdisk' utility support non-512 byte sized sector
sizes by allowing the user to specify the size the wish to use on the command
line. Update wic to provide this option in the case where a user wishes to
generate an image with a sector size other than 512 bytes.

Signed-off-by: Trevor Woerner <trevor.woerner@amd.com>
Signed-off-by: Mark Hatle <mark.hatle@amd.com>