xilinx-mirrors: Add new xilinx-mirrors.bbclass
The old way of using mirrors/premirrors to handle internal code development
is no longer working.  Instead we use the same concept of a mirror to adjust
the SRC_URI inline, but we directly tell the system to use our internal
locations.

Note, only one replacement is allowed.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>

Move to meta-xilinx-core layer.

Signed-off-by: Sandeep Gundlupet Raju <sandeep.gundlupet-raju@amd.com>
Signed-off-by: Mark Hatle <mark.hatle@amd.com>