meta-xilinx-core: xilinx-mirrors.bbclass: Add symlinks and fix hash calculation
Symlinks to the original filename(s) need to be generated so the output of
this can be used to seed a mirror for downstream users.

Additionally, we need to ensure that even though we are modifying the
SRC_URI, we do NOT want to change the recipe's calculated hash in any way.
Using code derived from bitbake's hash calculation, we can emulate
the correct behavior.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>