meta-xilinx-core: xilinx-mirrors.bbclass: Rework for consistent hash
Always ignore our variables __SRC_URI_MIRROR_LINKS mirror_fetcher_set_symlinks
when computing hashes.

The code was reorganized in the anonymous python.  It now sets the
vardepvalue for SRC_URI to ensure the value is static.  Unfortunately
we don't know if this could be an AMD specific repository or not,
so we have to apply this code to all repositories.

Note there may be corner cases where other anonymous python chunks may change
SRC_URI and the results won't be used in the hash calculation.

This class should not be used once the repositories information is
public, only during internal and EA releases.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>