xilinx-bootbin: Add post deploy, build dependency on extract-cdo
versal, versal-net and versal2 need to depend on extrat-cdo, or QEMU
will not be able to load if we do not have a ospi/qspi image.  QEMU requires
the pmc_cdo to be available, no ospi/qspi means that it can't extract it.

Note when booting without an ospi/qspi, you must also have your boot.bin in
the first (dos) partition of the disk image.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>