fwupd: Fix plugin_uefi_capsule dependencies
Override the packageconfig settings for plugin_uefi_capsule as per the below
diff compared to the original recipe:

-PACKAGECONFIG[plugin_uefi_capsule] = "-Dplugin_uefi_capsule=enabled -Dplugin_uefi_capsule_splash=false,-Dplugin_uefi_capsule=disabled,efivar fwupd-efi"
+PACKAGECONFIG[plugin_uefi_capsule] = "-Dplugin_uefi_capsule=enabled -Dplugin_uefi_capsule_splash=false -Defi_binary=false,-Dplugin_uefi_capsule=disabled,,fwupd-efi"

Explanation:

First we remove the efivar and fwupd-efi build dependencies as they are not
needed at build time. However, removing the fwupd-efi build dependency results
in this error:

| Run-time dependency fwupd-efi found: NO (tried pkgconfig and cmake)
| Looking for a fallback subproject for the dependency fwupd-efi
|
| ../fwupd-1.9.18/plugins/uefi-capsule/meson.build:118:15: ERROR: Automatic wrap-based subproject downloading is disabled

The error only occurs because the efi_binary option is set to true by default
which results in the build system trying to download and build the fwupd-efi
subproject and the download step is what produces the build error.

To fix this, disable the automatic download by setting efi_binary=false when
building plugin_uefi_capsule following yocto best practice. Since this binary is
only needed at run-time, instead add the fwupd-efi package that provides the
binary as RDEPENDS in the packageconfig (was DEPENDS).

Without the RDEPENDS, execution fails at run-time, for example:

$ FWUPD_UEFI_ESP_PATH=/boot/efi fwupdtool install /nfsroot/bootfw_capsule.cab
Loading?                 [*******************                    ]
WARNING: This package has not been validated, it may not work properly.
Waiting?                 [*******************                    ]
/usr/libexec/fwupd/efi/fwupdaa64.efi and /usr/libexec/fwupd/efi/fwupdaa64.efi.signed cannot be found

Now the fwupd-efi package is properly and automatically pulled into an image
build when enabling plugin_uefi_capsule.

Signed-off-by: Christian Kohn <chris.kohn@amd.com>
Signed-off-by: Mark Hatle <mark.hatle@amd.com>