bootbin: Add optional BIF_PARTITION_NAME variable
Add a new, optional BIF_PARTITION_NAME variable in bootbin and set
it to a default value in all versal* inc files. The new variable
allows the user to assign a custom name to an image id and populate
it in the generated bif file. For example

BIF_PARTITION_NAME[0x1c000000] ?= "apu_ss"

will set the name field of the image section with id 0x1c000000
to "apu_ss" as shown below

image {
    id = 0x1c000000, name=apu_ss
    { type=raw,  load=0x1000000, file=cortexa78-linux.dtb }
    { core=a78-0,  cluster=0,  exception_level = el-3,  trustzone, file=arm-trusted-firmware.elf }
    { core=a78-0,  cluster=0,  exception_level = el-2, file=u-boot.elf }
}

If multiple image ids are defined and assigned, then multiple
BIF_PARTITION_NAME[] values can be set, one for each respective id.

This new variable is optional, if not defined, the current behavior is
retained i.e. the default name "apu_ss" will be used.

The reason for introducing this new variable is so the user can assign
a custom name for each image whereas presently the name is hard-coded
and also, the same name ("apu_ss") is used for all images even though
they might have different ids and therefore should have different
names as well for better readability.

Signed-off-by: Christian Kohn <chris.kohn@amd.com>
Signed-off-by: Mark Hatle <mark.hatle@amd.com>