xilinx-bootbin: Remove dependency of boot.scr
A 'qemu-ospi/qspi.bin' is generated in order to fill in for the ospi or
qspi on a real board.  This file had a boot.scr appended to it in a
magic location, which was not being used.  This dependency caused issues
with the move to EFI.  Simply remove the dependency and usage of this
file.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>
(cherry picked from commit 3a337d180d2dd253f82aec593cb52386d7b834a6)
Signed-off-by: Mark Hatle <mark.hatle@amd.com>