wic: bump SRCREV and change branch
Change branch to "scarthgap" and update to the latest commit.

The non-scarthgap branch (no-create-var-sector), on master, requires an update
for poetry (the python build mechanism), therefore a scarthgap branch was
created to include the latest code, but without the poetry update.

The latest code update (SRCREV) includes the variable sector size plus the
help update.

Signed-off-by: Trevor Woerner <trevor.woerner@amd.com>
Signed-off-by: Mark Hatle <mark.hatle@amd.com>