packagegroup-openamp-fw-examples: restore original
The intent of this packagegroup was to use it to generate the build-time side
of these OpenAMP examples. When it was moved from an internal-only Yocto
layer to an external one, it was modified to act as a packagegroup for the
deploy-time side by mistake. This patch restores the original intent.

Fixes: 5e905ea84da6 ("meta-xilinx-demos: add OpenAMP components")
Signed-off-by: Trevor Woerner <trevor.woerner@amd.com>
Signed-off-by: Mark Hatle <mark.hatle@amd.com>