bootimg-efi-amd.py: Replace config file generation
Disable the bootloader config file generateion as they it will not work
for our configuration and there is no way to do this from within the WKS
file.

Add a find-and-replade loop which will search for the @@UUID@@ tag
within the systemd-boot and xen config files and replace it with the one
which is passed in the WKS sourceparams as "rootfs_uuid". This is done
to the temporary files which are staged for inclusion in the generated
wic file.

Signed-off-by: John Toomey <john.toomey@amd.com>