meta-xilinx-demos: openamp-fw-example-versal.inc: discover artifacts
Instead of hard-coding the example filenames and count for the FreeRTOS
vs Zephyr demos for Gen1 vs Gen2 before-hand, make the recipe more
flexible by having it process whatever it finds in the artifact bundles.
The artifact bundle names already include the OS, so that no longer
needs to be hard-coded either.

Simply take the artifacts as they are found in ${S} and place them in
/lib following the current dfx-mgr layout guidelines.

As a concrete example for the vek280, a bundle that contains:

	versal-vek280-sdt-seg_packagegroup-openamp-fw-examples/
	├── open-amp-xlnx-echo-freertos.elf
	├── open-amp-xlnx-matrix-freertos.elf
	└── open-amp-xlnx-proxy-freertos.elf

will get placed in the image as follows:

	/lib/firmware/xilinx/
	└── vek280
	    └── rpu
	        ├── vek280-r5-0-open-amp-xlnx-echo-freertos
	        │   └── vek280-r5-0-open-amp-xlnx-echo-freertos_slot0
	        │       └── vek280-r5-0-open-amp-xlnx-echo-freertos.elf
	        ├── vek280-r5-0-open-amp-xlnx-matrix-freertos
	        │   └── vek280-r5-0-open-amp-xlnx-matrix-freertos_slot0
	        │       └── vek280-r5-0-open-amp-xlnx-matrix-freertos.elf
	        └── vek280-r5-0-open-amp-xlnx-proxy-freertos
	            └── vek280-r5-0-open-amp-xlnx-proxy-freertos_slot0
	                └── vek280-r5-0-open-amp-xlnx-proxy-freertos.elf

Signed-off-by: Trevor Woerner <trevor.woerner@amd.com>
Message-ID: <20250929175355.368298-7-trevor.woerner@amd.com>
Signed-off-by: Mark Hatle <mark.hatle@amd.com>