dfx_user_dts: Add FW_INSTALL_DIR variable
Add a new FW_INSTALL_DIR variable and use that for the firmware
installation path instead of PN. Set the default FW_INSTALL_DIR
value to PN using weak assignment so the user can override it.
Then update all install commands to use the new FW_INSTALL_DIR
instead of PN. This patch does not introduce a functional
change but gives the user more flexibility to manipulate the
firmware installation directory from the firmware recipe that
inherits this class.

Signed-off-by: Christian Kohn <chris.kohn@amd.com>
Message-ID: <20250930213011.3909931-2-chris.kohn@amd.com>
Signed-off-by: Mark Hatle <mark.hatle@amd.com>