openamp-fw-example-versal.inc: check elf count
In this case the location of the folder containing the elf files of
OpenAMP demos changed, therefore this recipe was not able to deploy any
files. As written this recipe succeeded, but image creation later on
would end up failing. The strange thing is that both the old and the
new folders exist in ${WORKDIR} which kept this recipe from failing
outright. The old folder is still being created, but it is empty, and a
new folder has appeared that contains the firmware artifacts.

Add a check in this recipe to make sure at least one DEMO elf file is
found in ${S}, otherwise fail here instead of later.

Signed-off-by: Trevor Woerner <trevor.woerner@amd.com>
Signed-off-by: Mark Hatle <mark.hatle@amd.com>