xlnx-embeddedsw.bbclass: Name the embeddedsw to avoid conflict in SRCREV
When a recipe provides it's own application software that sits on top of
embeddedsw, the SRC_URI may contain both ESW and the application source.

This can result in a failure during parsing where SRCREV and SRCREV_app
may be set.  This combination is not valid, when there are multiple
git repositories, each must be named and have their own SRCREV defined.

Instead we switch to 'naming' the ESW component as 'embeddedsw' and
always configure SRCREV_embeddedsw to ensure applications can add
their own sources as necessary.

This resolves an issue that may presents itself similar to:

   ERROR:
   ../meta-xilinx/meta-xilinx-standalone-sdt/recipes-bsp/image-selector/image-selector_2024.2.bb:
   AUTOREV/SRCPV set too late for the fetcher to work properly, please set
   the variables earlier in parsing. Erroring instead of later obtuse build
   failures.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>
Message-Id: <20251016163254.4126512-2-mark.hatle@amd.com>
Signed-off-by: Sandeep Gundlupet Raju <sandeep.gundlupet-raju@amd.com>