meta-xilinx-standalone: embeddedsw-source: Run configure and patch before build
Normally do_build, the default task, ends up depending on a task that in
turn depends on other tasks all that way back to do_fetch.  One of these
tasks does the LIC_FILES_CHKSUM validation via a postfunc.

Due to this being a shared-work recipe, most tasks have been disabled,
which has the side effect of do_build never calling the task that verified
the license file checksum, nor calling do_configure, or do_patch.

In normal usage do_configure and do_patch are triggered prior to the copy
of the shared source, but this makes it difficult to diagnose if the
license may be specified incorrectly.

Add do_patch and do_configure before do_build to ensure the tasks run.
Also add the license validation as a postfunc of do_patch to ensure
that the license validation occurs.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>
Message-Id: <20251016163254.4126512-3-mark.hatle@amd.com>
Signed-off-by: Sandeep Gundlupet Raju <sandeep.gundlupet-raju@amd.com>