xlnx-embeddedsw.bbclass: Ensure the branch and variable values are hashed
The variable hashes may not always include items used in python style
expressions, such as "${@ ... }".  Introducing explicit variables and an
intermediate variable will ensure that the value is hashed and the inputs
themselves are also included in the hash.

The hashing will then in-turn cause the embeddedsw-source recipe to be
run, and new files generated even when minor changes are made, such as
the LIC_FILE_CHKSUM.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>
Message-Id: <20251016163254.4126512-4-mark.hatle@amd.com>
Signed-off-by: Sandeep Gundlupet Raju <sandeep.gundlupet-raju@amd.com>