machine: Fix QEMU issue with QB_DTB
A path relocation issue between qemuboot.bbclass and runqemu results
in the latter not working correctly if TMPDIR is not a subdirectory of
TOPDIR, for example when TMPDIR is set to /tmp/<some_dir>.

When the QEMU config fils is generated, a relative path is written if
the path starts with TOPDIR, otherwise an absolute path is used. runqemu
assumes that it is always going to be passed a relative path and
prefixes DEPLOY_DIR_IMAGE. The fallback behaviour is to glob for dtb
files which will likely result in QEMU not booting correctly.

This change removes DEPLOY_DIR_IMAGE from the QB_DTB variable result in
the correct relative path always being set in the QEMU config
(qemu-hw-devicetrees/multiarch/<board>.dtb) which runqemu can correctly
use in both scenarios.

Signed-off-by: John Toomey <john.toomey@amd.com>
Message-Id: <20251016190134.2091946-1-john.toomey@amd.com>
Signed-off-by: Sandeep Gundlupet Raju <sandeep.gundlupet-raju@amd.com>