meta-xilinx-mali400: Add additional QT related software to machine_arch
When QT software is build using libmali, many things that use QT may end up
being configured to use libmali instead of mesa.  Add a number of items
from both meta-qt5 and meta-ros to this set, ensuring that the
dynamic-layer is used to control when the bbappend is available.

This uses a standard formula of capturing the original recipe package_arch,
checking if libmali is enabled to set the machine_arch, and then
selecting between the package_arch and machine_arch based on a recipe
specific dependency that introduces the behavior.  In many cases
this will be qtbase, but other may have indirectly dependencies
on qtbase that end up contaminating their configurations.

The easiest way to find this is to run two builds, one with the
standard mesa configuration, and one with the libmali configuration
comparing the resulting packages to identify contamination.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>
Message-Id: <20251021021902.3105807-4-mark.hatle@amd.com>
Signed-off-by: Sandeep Gundlupet Raju <sandeep.gundlupet-raju@amd.com>