meta-xilinx-core: machines: Update configuration to use KERNEL_IMAGETYPE
Instead of using a hardcoded kernel type, use the KERNEL_IMAGETYPE as
defined by the standard Yocto Project usage.  This will allow the user
to specify any valid kernel type for a given architecture.  The arch
defaults remain as before the change, uImage for Zynq and Image for
zynqmp, and versal family.

This adds the ability to change KERNEL_IMAGETYPE to a compressed
kernel, such as "Image.lzma" which dramatically reduces the size
of the Linux kernel in the storage. (30+ MB to 4-ish MB in some
configurations.)

Signed-off-by: Mark Hatle <mark.hatle@amd.com>
Message-ID: <20251205174016.3639977-2-mark.hatle@amd.com>
Signed-off-by: Mark Hatle <mark.hatle@amd.com>