meta-xilinx-imgrcvry: image-recovery-tiny-initramfs: Move IMAGE setting into recipe
The settings IMAGE_BOOT_FILES and EXTRA_IMAGEDEPENDS are defaulted by the
machine, but ultimately it is up to the image recipe to do the correct
thing with these values.  Move the setting from the distribution to the
image recipe so other image recipes that may be used are not
impacted.

KERNEL_DEPLOY_DEPEND is an existing setting from the image.bbclass.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>
Message-ID: <20251208212825.1366184-6-mark.hatle@amd.com>
Signed-off-by: Mark Hatle <mark.hatle@amd.com>